feat: publishing all events to product api through event handler and kafka

+++ src/buckstore.manager.service.application/Commands/UpdateProductCommand.cs
@@ -9,10 +9,10 @@ namespace buckstore.manager.service.application.Commands
         public Guid ProductCode { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
         public int Category { get; set; }
-        
+
         public override bool IsValid()
         {
             ValidationResult = new UpdateProductValidations().Validate(this);
